Don't sleep in wapbl_flush() if we are not supposed to.
wapbl_flush() called with waitfor = 0 is expected to perform an
opportunistic, best-effort flush of the journal; it is not meant to
sleep for locks.

Please note that wapbl_flush() should only be called with waitfor = 1
from places where we know no vnode locks are being held, otherwise
deadlocks might occur.